refactor(cli): separate parser data and main function logic
